left canonical transformation and appropriaate tests are added
